separate tns retrieval and submission queues
